+++ lib/polars/string_name_space.rb
@@ -1494,13 +1494,7 @@ module Polars
     #   # [
     #   #         null
     #   # ]
-    def join(delimiter = nil, ignore_nulls: true)
-      # TODO update
-      if delimiter.nil?
-        warn "The default `delimiter` for `join` method will change from `-` to empty string in a future version"
-        delimiter = "-"
-      end
-
+    def join(delimiter = "", ignore_nulls: true)
       super
     end